% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard_extractor.R
\name{extract_chart}
\alias{extract_chart}
\title{Extract information from Billboard chart}
\usage{
extract_chart(url)
}
\arguments{
\item{url}{A Billboard URL that points to a chart}
}
\value{
A list whoose entries are the rows of the extracted billboard chart
}
\description{
Given a Billboard chart URL, extracts the chart information
to a list
}
