% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billboard_extractor.R
\name{hot100}
\alias{hot100}
\title{Billboard Hot 100 chart}
\usage{
hot100(top_date)
}
\arguments{
\item{top_date}{A Date object or a string in YYYY-DD-MM format}
}
\value{
A list whoose entries are the billboard chart
}
\description{
Returns the Billboard Hot 100 chart from a specifed date
}
